+++ Grpc.AspNetCore.sln
@@ -7,8 +7,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{EA3305
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{8C62055F-8CD7-4859-9001-634D544DF2AE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grpc.Core.Api", "src\Grpc.Common\Grpc.Core.Api.csproj", "{A06B88C7-8CD1-4C5F-9DE7-5D66A252DAED}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Proto", "Proto", "{AC41B2B2-62D9-454E-8521-F71E331B07F3}"
 	ProjectSection(SolutionItems) = preProject
 		sample\Proto\chat.proto = sample\Proto\chat.proto
@@ -43,15 +41,16 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Certs", "Certs", "{409A27EF
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{3E045DF6-62EF-4331-B712-9CF5EC56F04C}"
 	ProjectSection(SolutionItems) = preProject
-		sample\Shared\ClientResources.cs = sample\Shared\ClientResources.cs
-		sample\Shared\Resources.cs = sample\Shared\Resources.cs
+		sample\Shared\Utils.cs = sample\Shared\Utils.cs
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grpc.AspNetCore", "src\Grpc.AspNetCore\Grpc.AspNetCore.csproj", "{5E6197EE-9298-4922-B612-D6D933AB6164}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{CECC4AE8-9C4E-4727-939B-517CC2E58D65}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{20F5CF21-E075-4B5E-AFF4-CECD6561D524}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grpc.AspNetCore.Server", "src\Grpc.AspNetCore.Server\Grpc.AspNetCore.Server.csproj", "{89ED416F-92F1-4425-9379-D4E76A285860}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grpc.AspNetCore.Tests", "test\Grpc.AspNetCore.Tests\Grpc.AspNetCore.Tests.csproj", "{B721E219-5A43-4334-B1A3-A8B1D9E498E6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grpc.Core.Api", "src\Grpc.Core.Api\Grpc.Core.Api.csproj", "{7AB17083-D5F5-4B19-B054-EC191001A80D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grpc.AspNetCore.Server.Tests", "test\Grpc.AspNetCore.Server.Tests\Grpc.AspNetCore.Server.Tests.csproj", "{278B1311-41BF-4345-947E-7B70E277FC3C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,10 +58,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A06B88C7-8CD1-4C5F-9DE7-5D66A252DAED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A06B88C7-8CD1-4C5F-9DE7-5D66A252DAED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A06B88C7-8CD1-4C5F-9DE7-5D66A252DAED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A06B88C7-8CD1-4C5F-9DE7-5D66A252DAED}.Release|Any CPU.Build.0 = Release|Any CPU
 		{47795F46-3E85-4B13-85E3-3344555600F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{47795F46-3E85-4B13-85E3-3344555600F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{47795F46-3E85-4B13-85E3-3344555600F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -79,20 +74,23 @@ Global
 		{1E882855-D205-4139-BE55-C44511FF43B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1E882855-D205-4139-BE55-C44511FF43B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1E882855-D205-4139-BE55-C44511FF43B0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5E6197EE-9298-4922-B612-D6D933AB6164}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5E6197EE-9298-4922-B612-D6D933AB6164}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5E6197EE-9298-4922-B612-D6D933AB6164}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5E6197EE-9298-4922-B612-D6D933AB6164}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B721E219-5A43-4334-B1A3-A8B1D9E498E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B721E219-5A43-4334-B1A3-A8B1D9E498E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B721E219-5A43-4334-B1A3-A8B1D9E498E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B721E219-5A43-4334-B1A3-A8B1D9E498E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{89ED416F-92F1-4425-9379-D4E76A285860}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89ED416F-92F1-4425-9379-D4E76A285860}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89ED416F-92F1-4425-9379-D4E76A285860}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89ED416F-92F1-4425-9379-D4E76A285860}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7AB17083-D5F5-4B19-B054-EC191001A80D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7AB17083-D5F5-4B19-B054-EC191001A80D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7AB17083-D5F5-4B19-B054-EC191001A80D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7AB17083-D5F5-4B19-B054-EC191001A80D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{278B1311-41BF-4345-947E-7B70E277FC3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{278B1311-41BF-4345-947E-7B70E277FC3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{278B1311-41BF-4345-947E-7B70E277FC3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{278B1311-41BF-4345-947E-7B70E277FC3C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{A06B88C7-8CD1-4C5F-9DE7-5D66A252DAED} = {8C62055F-8CD7-4859-9001-634D544DF2AE}
 		{AC41B2B2-62D9-454E-8521-F71E331B07F3} = {EA330592-E294-4A29-B60E-196AA05831DD}
 		{F6E0F9D7-64E5-4C7B-A9BC-3C2AD687710B} = {EA330592-E294-4A29-B60E-196AA05831DD}
 		{A7E09CBD-853F-4C42-B4C4-A1C2CBA10C0B} = {EA330592-E294-4A29-B60E-196AA05831DD}
@@ -102,8 +100,9 @@ Global
 		{1E882855-D205-4139-BE55-C44511FF43B0} = {F6E0F9D7-64E5-4C7B-A9BC-3C2AD687710B}
 		{409A27EF-A5E9-45C5-9041-7210C6C8FC17} = {EA330592-E294-4A29-B60E-196AA05831DD}
 		{3E045DF6-62EF-4331-B712-9CF5EC56F04C} = {EA330592-E294-4A29-B60E-196AA05831DD}
-		{5E6197EE-9298-4922-B612-D6D933AB6164} = {8C62055F-8CD7-4859-9001-634D544DF2AE}
-		{B721E219-5A43-4334-B1A3-A8B1D9E498E6} = {20F5CF21-E075-4B5E-AFF4-CECD6561D524}
+		{89ED416F-92F1-4425-9379-D4E76A285860} = {8C62055F-8CD7-4859-9001-634D544DF2AE}
+		{7AB17083-D5F5-4B19-B054-EC191001A80D} = {8C62055F-8CD7-4859-9001-634D544DF2AE}
+		{278B1311-41BF-4345-947E-7B70E277FC3C} = {CECC4AE8-9C4E-4727-939B-517CC2E58D65}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CD5C2B19-49B4-480A-990C-36D98A719B07}
