+++ sample/Server/Server.csproj
@@ -22,7 +22,7 @@
 
     <Compile Include="$(MSBuildThisFileDirectory)..\Shared\Resources.cs" />
 
-    <ProjectReference Include="../../src/Grpc.AspNetCore/Grpc.AspNetCore.csproj" />
+    <ProjectReference Include="..\..\src\Grpc.AspNetCore.Server\Grpc.AspNetCore.Server.csproj" />
 
     <PackageReference Include="Microsoft.AspNetCore.App" Version="3.0.0-preview-18579-0056" />
     <PackageReference Include="Grpc.Tools" Version="1.17.0-pre3" />
