+++ src/Grpc.AspNetCore.Server/GrpcServiceExtensions.cs
@@ -17,7 +17,8 @@
 #endregion
 
 using System;
-using Grpc.AspNetCore.Internal;
+using Grpc.AspNetCore.Server;
+using Grpc.AspNetCore.Server.Internal;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace Microsoft.Extensions.DependencyInjection
